pyproject.toml
[tool.poetry]
name = "oddify"
version = "0.1.0"
description = "Pokemon... But Oddish???"
authors = ["Liam Ilan <snowboardsheep@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
discord = "^1.0.1"
python-dotenv = "^0.17.1"
Flask = "^2.0.1"
gunicorn = "^20.1.0"
Pillow = "^8.2.0"
requests = "^2.25.1"
pycountry = "^20.7.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
